Add client-side form validation to form in 'about' section with Javascript.
* [Style] - Added styles for :valid, :invalid, :invalid:required on inputs for full name and emails.
*[Style] - Added styles for .error and .active classes for the error messages upon invalid or incomplete entries in the 'full name' and 'email' fields of the form.
[HTML] - Added error class to 'full name' and 'email' address inputs
[Accessibility] - Added Aria-live = 'polite' to 'full name' and 'email' span tags used to print the errors.

Changes will prevent be better conformed to accessibility guidelines which requires required fields to be advised and will allow users to get a polite notification from their screen reader if the leave a required field blank.

Changes also provide visual cues that do not interfere with the design of the page.