Add contact form without backend code
Users were unable to submit information through the contact form in the 'About' section because there was not a backend server implemented due to not having the required knowledge or experience to do so. As a solution, the 'Formkeep' forms was added as the action attribute with a method of POST.

POST method will prevent user data from being shown in the browser when the form is sent. Form keep was chosen because of being able to send notifications to email once a day with messages received. Form keep allows 50 submissions per month.

It is worthy to note that depending on how much interest the profile generates, a paid service may become necessary to unlock more features. Once a custom domain is decided upon, the url address in the Form Settings > General section of 'Formkeep' will need updated to match the new custom URL.