Add Screenshot of Desktop View
The project cards were not scaling properly because viewport units were being used. A max width in ems was added to prevent the cards from getting too big or too small. Cards will not grow up to 25em for mobile and at 1200px the lay out will be able to grow up to 75em.

[Documentation] Updated README.md to explain why the grid system was used in the desktop layout and the updates that were added to version 1.0.1