[Style] Make page more responsive on mobile devices
links for live view and code view were not center aligned in mobile because they further nested into a list element and flex properties were not properly applied to the list.

contact form was too narrow on mobile devices because the wrong measurement unit was used for the min-width property. removed the min-width property from project cards and project section.
Links for twitter and FEM were did not have enough contrast against the white background. Changed the text color from yellow to the same color as the main text color.
Increased the spacing between the about section and the rest of the content by giving each section the min-height: 100% property.
The contact form is no longer too small on mobile, but it is still not sufficiently responsive as it overflows the screen and gets cut off.
Added aria-labelled by to each section

[Documentation] Added link to Kevin Powell's sections are a lie (sort of) video.