[Style] Center align web content on phone screen
content was left-aligned on devices 1080px X 2400px consistent with the 'Galaxy Fold' view in the dev tool. The form was cut off on the left side and the entire page was unpleasant to look at.

removed 100vh from 'About Me' sections and ultimately realized that the problem was with some of the content not inheriting the flexCenter mixin from the body since it was nested inside other sections. Adjusted CSS rule in the main.scss file so that the rules for the body would also applied to main and main's direct descendants.

It is worthy to note that since the issue appears to be fixed, 100vh may be added back into the 'About' Section.