Improve accessibility on mobile view
The contact form lacked padding which made it more difficult to read the labels. Added padding of 2em on all 4 sides. However, this caused the contact form to overflow the viewport on mobile devices, so a max-width of 90vw was added to the form. This fixed the issue.

The size of the navigation bar would increase on hover, making it seem as if the navigation bar was 'jumping' on hover. This was due to the border being added back to the list items in navigation after having been originally removed for easier customization when the page was first created. This commit adds a transparent border back to the navigation links, which then turns yellow on hover. This resolves the issue of the navigation 'jumping' around on hover.

The project cards were too narrow on mobile devices. Each .card originally had a margin of 4em on the top and bottom and 2em on the left and right sides. Chaging the margin to auto for the left and right sides corrected the issue. Other fixes that were tried were removing the margin from the .container__projects, removing flex alignment rules for the section (justify-content: center, align-items: center, text-align: center). It is worthy to note that although removing these styles did not correct the issue, they did not affect the layout at all, so it was best to remove them since they were not needed.