Correct link alignment for project cards
Links in the project cards were not centered aligned in mobile view. This is because there was padding only on the left side of the <ul> tag, which was pushing the links to the right. Adding padding on the remaining sides aligned the links. Other solutions tried were removing the entirely, but the links were still pushed to the right. Adding another line of code to add padding just on the right and left would have worked, but it was adding padding on all sides was shorter code with the same result.

The tag line in the header was reduced from 'I am a Front End Web Developer committed to bringing your projects to life' to 'Front End Web Developer' it was crowded in the mobile view and it was repeated in the 'About Me.'

The 'About Me' was shortened to highlight the value brought to the client more succintly. Removed references to level of education, and generic 'web experiences.'