Remove padding from <ul> in project cards
Previous fix of adding padding on all sides did not work on mobile, but only noticeable once it was pushed, as per last commit. Removed padding from the ul in project card explicitly and added a separate rule set for desktop view for these links because when the links looked like they were center aligned in mobile, the would not be in desktop view.