Add Illustration to Header
User believed Header was too plain on desktop, so the Proud Coder illustration was added from Undraw.co. on devices with tablet size screens and larger.

User became frustrated with not having a way to navigate back to the top of the page without having to travel the entire screen with mouse or finger. A back to top arrow link was added at the bottom of the page.