Path to hero image not working
SVG image in hero is not working when pushed to Github Pages, but it worked locally. The after pseudo-element is being added to HTML, but the image is not loading. Tried editing the path in browser by adding a '.' at the beginning of the relative path, but that did not work. Tried adding the SVG as a background image to the pseudo-element but that also did not work. Tried adding the svg directly as content of the pseudo-element, that did not work. Tried adding the svg as an image element in the HTML, and although that would work, the image should only be displayed screens that are at least 40em wide.