Thumbnail preview of shared link did not update
The path was incorrect for the image, so the preview of image of the link when shared did not update accordingly.

Using the sharing debugger on meta for developers, it was determined that the 'og: image' meta should be explicitly provided, even if a value can be inferred from other tags.

Took a better screenshot of what the thumbnail should be and provided the explicit url of the image. Will test once pushed live but prior to sharing.