preview thumbnail of shared link still not working
The thumbnail preview of the shared link is still not updated even after explicity providing the image url. The sharing debugger continues to scrape the three-column-preview project as the thumbnail of the entire page, which is not what should be happening.

The 'og:image' meta tag has been moved up to come right after the opening head tag. Will retry.